Add migrator events

+++ download.py
@@ -38,9 +38,10 @@ def download_from_url(url: str, base_path: str, filename: str=None, retry: int=0
 
 def run_post_scripts(assets: dict):
     post_scripts: dict = assets.get('post_scripts', {})
-    apply_hard_links(assets['target_path'], post_scripts.get('hard_links'))
-    apply_actionscript_patches(assets['target_path'], post_scripts.get('actionscript_patches'))
-    apply_swf_replacements(assets['target_path'], post_scripts.get('swf_replacements'))
+    apply_hard_links(assets['target_path'], post_scripts.get('hard_links', []))
+    apply_actionscript_patches(assets['target_path'], post_scripts.get('actionscript_patches', []))
+    apply_swf_replacements(assets['target_path'], post_scripts.get('swf_replacements', []))
+    apply_swf_removals(assets['target_path'], post_scripts.get('swf_removals', []))
     
 def apply_hard_links(base_path: str, hard_links: list):
     for [hard_link_target, hard_link_name] in hard_links:
@@ -67,7 +68,8 @@ def apply_actionscript_patches(base_path: str, as_patches: list[dict]):
             ["./jpexs/replace_actionscript.sh", path, "DoAction.as"] + ([script_path_target] if script_path_target else []),
             check=True, capture_output=True)
         os.replace("./output.swf", path)
-    os.remove("./DoAction.as")
+    if os.path.exists("./DoAction.as"):
+        os.remove("./DoAction.as")
     
 def apply_swf_replacements(base_path: str, swf_replaces: list[dict]):
     for item in swf_replaces:
@@ -78,6 +80,15 @@ def apply_swf_replacements(base_path: str, swf_replaces: list[dict]):
                 ["./jpexs/replace_asset.sh", path, replacement["asset_filename"], str(replacement["id"])] ,
             check=True, capture_output=True)
             os.replace("./output.swf", path)
+            
+def apply_swf_removals(base_path: str, swf_removals: list[dict]):
+    for item in swf_removals:
+        print(f'Removing assets for {item["swf_filename"]}')
+        path = os.path.join(base_path, item["swf_filename"])
+        subprocess.run(
+            ["./jpexs/remove_asset.sh", path, " ".join([str(x) for x in item["removals"]])],
+            check=True, capture_output=True)
+        os.replace("./output.swf", path)
 
 def main(assets_source_file):
     with open(assets_source_file) as f:
